Finish graphics processing code (except testing)
The graphics code can clear the screen, and draw to it. Currently, there
is no window on which to draw the pixels; only the array which contains
the pixel values is filled.
Sprites have their origin at the top left corner, and each sprite is 8
bits wide, and n bytes tall. They are also read from the highest bit,
which represents the top left corner.
When a pixel is turned from 1 to 0 during the draw, the VF register is
set, otherwise it is cleared. This is represented by returning a boolean
value from the draw() function.