Add `Graphics` trait and update to latest Rust
This project hasn't been updated in years, so the dependencies were
bumped to the newest versions. Changed APIs were updated to match. The
SDL dependency builds with VS 17 (2022) now as well.

`Graphics` is renamed to `GraphicsImpl` and `Graphics` is now a trait.
This will help with wiring up the components for the whole emulator. A
`screen()` method was added so that the contents of the screen can be
retrieved.